# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/pricing-matrices": {
            "get": {
                "description": "Get paginated list of pricing matrices with optional filters for customer, origin city, destination city, and status",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pricing_matrix"
                ],
                "summary": "Get list of pricing matrices",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Filter by customer ID",
                        "name": "customer_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by origin city ID",
                        "name": "origin_city_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by destination city ID",
                        "name": "destination_city_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Filter by status (active/inactive)",
                        "name": "status",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Number of items per page",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "format": "int64",
                        "description": "Page number",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Search term",
                        "name": "search",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Order by field (id/-id)",
                        "name": "order_by",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bearer jwt-token...",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new pricing matrix with customer, route, vehicle type, and pricing details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pricing_matrix"
                ],
                "summary": "Create a new pricing matrix",
                "parameters": [
                    {
                        "description": "Pricing matrix data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pricing_matrix.createRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer jwt-token...",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    }
                }
            }
        },
        "/pricing-matrices/{id}": {
            "get": {
                "description": "Get detailed information about a specific pricing matrix by ID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pricing_matrix"
                ],
                "summary": "Get pricing matrix details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pricing matrix ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer jwt-token...",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    }
                }
            },
            "put": {
                "description": "Update an existing pricing matrix with new customer, route, vehicle type, or pricing details",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pricing_matrix"
                ],
                "summary": "Update a pricing matrix",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pricing matrix ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Pricing matrix data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/pricing_matrix.updateRequest"
                        }
                    },
                    {
                        "type": "string",
                        "description": "Bearer jwt-token...",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "description": "Soft delete a pricing matrix by ID (marks as deleted without removing from database)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pricing_matrix"
                ],
                "summary": "Delete a pricing matrix",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pricing matrix ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer jwt-token...",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    }
                }
            }
        },
        "/pricing-matrices/{id}/activate": {
            "put": {
                "description": "Activate a pricing matrix to make it available for use in pricing calculations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pricing_matrix"
                ],
                "summary": "Activate a pricing matrix",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pricing matrix ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer jwt-token...",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    }
                }
            }
        },
        "/pricing-matrices/{id}/deactivate": {
            "put": {
                "description": "Deactivate a pricing matrix to prevent it from being used in pricing calculations",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "pricing_matrix"
                ],
                "summary": "Deactivate a pricing matrix",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Pricing matrix ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer jwt-token...",
                        "name": "authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/rest.ResponseBody"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/rest.HTTPError"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "pricing_matrix.createRequest": {
            "type": "object",
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "destination_city_id": {
                    "type": "string"
                },
                "origin_city_id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
        "pricing_matrix.updateRequest": {
            "type": "object",
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "destination_city_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "origin_city_id": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        },
...
        "rest.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "$ref": "#/definitions/rest.Message"
                }
            }
        },
        "rest.Message": {
            "type": "string",
            "enum": [
                "success",
                "resource created",
                "resource updated",
                "resource deleted",
                "invalid request body",
                "missing required fields",
                "invalid field value",
                "validation failed",
                "unauthorized",
                "forbidden",
                "resource not found",
                "conflict",
                "internal server error",
                "service unavailable",
                "invalid request body. please check your input format",
                "method not allowed"
            ],
            "x-enum-varnames": [
                "MsgSuccess",
                "MsgCreated",
                "MsgUpdated",
                "MsgDeleted",
                "MsgInvalidJSON",
                "MsgMissingField",
                "MsgInvalidField",
                "MsgValidationError",
                "MsgUnauthorized",
                "MsgForbidden",
                "MsgNotFound",
                "MsgConflict",
                "MsgInternalError",
                "MsgServiceUnavailable",
                "MsgBadRequest",
                "MsgNotAllowed"
            ]
        },
        "rest.Meta": {
            "type": "object",
            "properties": {
                "has_next": {
                    "type": "boolean"
                },
                "has_prev": {
                    "type": "boolean"
                },
                "page": {
                    "type": "integer"
                },
                "page_size": {
                    "type": "integer"
                },
                "total": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
        "rest.ResponseBody": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "meta": {
                    "$ref": "#/definitions/rest.Meta"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
    }
}